refactor: extract error constants
