+++ src/routes/Notifications.test.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 import { mockAccountNotifications } from '../__mocks__/notifications-mocks';
 import { mockSettings } from '../__mocks__/state-mocks';
 import { AppContext } from '../context/App';
-import { Errors } from '../utils/constants';
+import { Errors } from '../utils/errors';
 import { NotificationsRoute } from './Notifications';
 
 jest.mock('../components/AccountNotifications', () => ({
